Add Some Custom Exception

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType
@@ -17,6 +18,7 @@ namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeave
             //Convert to domain entity object
             var leaveTypeDelete = await this._leaveTypeRepository.GetByIdAsync(request.Id);
             //verify record exist
+            if (leaveTypeDelete == null) throw new NotFoundException(nameof(LeaveType),request.Id);
             //Delete to Database
             await this._leaveTypeRepository.DeleteAsync(leaveTypeDelete);
             //return record Id
